packages/django-layers/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "django-layers"
version = "0.1.0"
description = "Import boundary enforcement for Django monorepos"
readme = "README.md"
license = {text = "Proprietary"}
requires-python = ">=3.11"
authors = [
    {name = "Nestor Wheelock", email = "nestor@wheelock.dev"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: Other/Proprietary License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Quality Assurance",
]
keywords = ["django", "architecture", "layers", "imports", "linter", "ci"]
dependencies = [
    "PyYAML>=6.0",
]

[project.urls]
Homepage = "https://github.com/nestorwheelock/django-layers"
Repository = "https://github.com/nestorwheelock/django-layers"

[project.optional-dependencies]
dev = [
    "pytest",
]

[project.scripts]
django-layers = "django_layers.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
testpaths = ["tests"]
pythonpath = ["src", "."]
addopts = "-v --tb=short"
norecursedirs = ["fixtures"]
